feat: migrate news.ts to typescript
