+++ .eslintrc copy.json
@@ -19,3 +19,7 @@
         "class-methods-use-this": "off"
     }
 }
+
+
+/*         // eslint . --ext .ts
+ */
