+++ package.json
@@ -1,43 +1,42 @@
 {
-    "name": "news-js",
-    "version": "1.0.0",
-    "description": "",
-    "main": "webpack.base.config.js",
-    "scripts": {
-        "start": "webpack serve --open --config ./webpack.config.js --env mode=dev",
-        "build": "webpack --config ./webpack.config.js --env mode=prod",
-        "lint": "eslint src --ext .ts"
-        // eslint . --ext .ts
-    },
-    "repository": {
-        "type": "git",
-        "url": ""
-    },
-    "keywords": [],
-    "author": "Aleh Serhiyenia",
-    "license": "ISC",
-    "homepage": "",
-    "devDependencies": {
-        "@typescript-eslint/eslint-plugin": "^5.5.0",
-        "@typescript-eslint/parser": "^5.5.0",
-        "clean-webpack-plugin": "^3.0.0",
-        "copy-webpack-plugin": "^7.0.0",
-        "css-loader": "^5.1.0",
-        "eslint": "^7.32.0",
-        "eslint-config-airbnb-typescript": "^16.1.0",
-        "eslint-config-prettier": "^8.3.0",
-        "eslint-plugin-import": "^2.25.3",
-        "eslint-plugin-prettier": "^3.4.0",
-        "eslint-webpack-plugin": "^3.1.1",
-        "html-webpack-plugin": "^5.2.0",
-        "prettier": "2.2.1",
-        "style-loader": "^2.0.0",
-        "ts-loader": "^9.2.6",
-        "typescript": "^4.5.2",
-        "typescript-eslint-parser": "^22.0.0",
-        "webpack": "^5.37.1",
-        "webpack-cli": "^4.5.0",
-        "webpack-dev-server": "^3.11.2",
-        "webpack-merge": "^5.7.3"
-    }
+  "name": "news-js",
+  "version": "1.0.0",
+  "description": "",
+  "main": "webpack.base.config.js",
+  "scripts": {
+    "start": "webpack serve --open --config ./webpack.config.js --env mode=dev",
+    "build": "webpack --config ./webpack.config.js --env mode=prod",
+    "lint": "eslint src --ext .ts"
+  },
+  "repository": {
+    "type": "git",
+    "url": ""
+  },
+  "keywords": [],
+  "author": "Aleh Serhiyenia",
+  "license": "ISC",
+  "homepage": "",
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.5.0",
+    "@typescript-eslint/parser": "^5.5.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^7.0.0",
+    "css-loader": "^5.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-typescript": "^16.1.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-webpack-plugin": "^3.1.1",
+    "html-webpack-plugin": "^5.2.0",
+    "prettier": "2.2.1",
+    "style-loader": "^2.0.0",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.5.2",
+    "typescript-eslint-parser": "^22.0.0",
+    "webpack": "^5.37.1",
+    "webpack-cli": "^4.5.0",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.7.3"
+  }
 }
