+++ src/components/view/news/news.css
@@ -1,166 +1,166 @@
 .news__item {
-    display: flex;
-    flex-direction: column;
-    margin: 1rem auto;
-    margin-bottom: 1.6%;
-    background: #fff;
-    color: #333;
-    line-height: 1.4;
-    font-family: Arial, sans-serif;
-    border-radius: 5px;
-    overflow: hidden;
+  display: flex;
+  flex-direction: column;
+  margin: 1rem auto;
+  margin-bottom: 1.6%;
+  background: #fff;
+  color: #333;
+  line-height: 1.4;
+  font-family: Arial, sans-serif;
+  border-radius: 5px;
+  overflow: hidden;
 }
 
 .news__item:hover .news__meta-photo {
-    transform: scale(1.3) rotate(3deg);
+  transform: scale(1.3) rotate(3deg);
 }
 
 .news__item .news__meta {
-    position: relative;
-    height: 200px;
+  position: relative;
+  height: 200px;
 }
 
 .news__item .news__meta-photo {
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    background-size: cover;
-    background-position: center;
-    transition: transform 0.2s;
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  background-size: cover;
+  background-position: center;
+  transition: transform 0.2s;
 }
 
 .news__item .news__meta-details,
 .news__item .news__meta-details ul {
-    margin: auto;
-    padding: 0;
-    list-style: none;
+  margin: auto;
+  padding: 0;
+  list-style: none;
 }
 
 .news__item .news__meta-details {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: -120%;
-    margin: auto;
-    transition: left 0.2s;
-    background: rgba(0, 0, 0, 0.6);
-    color: #fff;
-    padding: 10px;
-    width: 100%;
-    font-size: 0.9rem;
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: -120%;
+  margin: auto;
+  transition: left 0.2s;
+  background: rgba(0, 0, 0, 0.6);
+  color: #fff;
+  padding: 10px;
+  width: 100%;
+  font-size: 0.9rem;
 }
 
 .news__item .news__description {
-    padding: 1rem;
-    background: #fff;
-    position: relative;
-    z-index: 1;
+  padding: 1rem;
+  background: #fff;
+  position: relative;
+  z-index: 1;
 }
 
 .news__item .news__description h2 {
-    line-height: 1;
-    margin: 0;
-    font-size: 1.7rem;
+  line-height: 1;
+  margin: 0;
+  font-size: 1.7rem;
 }
 
 .news__item .news__description h3 {
-    font-size: 1rem;
-    font-weight: 300;
-    text-transform: uppercase;
-    color: #a2a2a2;
-    margin-top: 5px;
+  font-size: 1rem;
+  font-weight: 300;
+  text-transform: uppercase;
+  color: #a2a2a2;
+  margin-top: 5px;
 }
 
 .news__item .news__description .news__read-more {
-    text-align: right;
+  text-align: right;
 }
 
 .news__item .news__description .news__read-more a {
-    color: #5ad67d;
-    display: inline-block;
-    position: relative;
-    text-decoration: none;
-    font-weight: 800;
+  color: #5ad67d;
+  display: inline-block;
+  position: relative;
+  text-decoration: none;
+  font-weight: 800;
 }
 
 .news__item .news__description .news__read-more a:after {
-    content: '→';
-    margin-left: -10px;
-    opacity: 0;
-    vertical-align: middle;
-    transition: margin 0.3s, opacity 0.3s;
+  content: '→';
+  margin-left: -10px;
+  opacity: 0;
+  vertical-align: middle;
+  transition: margin 0.3s, opacity 0.3s;
 }
 
 .news__item .news__description .news__read-more a:hover:after {
-    margin-left: 5px;
-    opacity: 1;
+  margin-left: 5px;
+  opacity: 1;
 }
 
 .news__item p {
-    margin: 1rem 0 0;
+  margin: 1rem 0 0;
 }
 
 .news__item p:first-of-type {
-    margin-top: 1.25rem;
-    position: relative;
+  margin-top: 1.25rem;
+  position: relative;
 }
 
 .news__item p:first-of-type:before {
-    content: '';
-    position: absolute;
-    height: 5px;
-    background: #5ad67d;
-    width: 35px;
-    top: -0.75rem;
-    border-radius: 3px;
+  content: '';
+  position: absolute;
+  height: 5px;
+  background: #5ad67d;
+  width: 35px;
+  top: -0.75rem;
+  border-radius: 3px;
 }
 
 .news__item:hover .news__meta-details {
-    left: 0%;
+  left: 0%;
 }
 
 @media (min-width: 640px) {
-    .news__item {
-        flex-direction: row;
-        max-width: 700px;
-    }
+  .news__item {
+    flex-direction: row;
+    max-width: 700px;
+  }
 
-    .news__item .news__meta {
-        flex-basis: 40%;
-        height: auto;
-    }
+  .news__item .news__meta {
+    flex-basis: 40%;
+    height: auto;
+  }
 
-    .news__item .news__description {
-        flex-basis: 60%;
-    }
+  .news__item .news__description {
+    flex-basis: 60%;
+  }
 
-    .news__item .news__description:before {
-        -webkit-transform: skewX(-3deg);
-        transform: skewX(-3deg);
-        content: '';
-        background: #fff;
-        width: 30px;
-        position: absolute;
-        left: -10px;
-        top: 0;
-        bottom: 0;
-        z-index: -1;
-    }
+  .news__item .news__description:before {
+    -webkit-transform: skewX(-3deg);
+    transform: skewX(-3deg);
+    content: '';
+    background: #fff;
+    width: 30px;
+    position: absolute;
+    left: -10px;
+    top: 0;
+    bottom: 0;
+    z-index: -1;
+  }
 
-    .news__item.alt {
-        flex-direction: row-reverse;
-    }
+  .news__item.alt {
+    flex-direction: row-reverse;
+  }
 
-    .news__item.alt .news__description:before {
-        left: inherit;
-        right: -10px;
-        -webkit-transform: skew(3deg);
-        transform: skew(3deg);
-    }
+  .news__item.alt .news__description:before {
+    left: inherit;
+    right: -10px;
+    -webkit-transform: skew(3deg);
+    transform: skew(3deg);
+  }
 
-    .news__item.alt .news__meta-details {
-        padding-left: 25px;
-    }
+  .news__item.alt .news__meta-details {
+    padding-left: 25px;
+  }
 }
