+++ src/index.ts
@@ -2,5 +2,5 @@ const App = require('./components/app/app');
 // import App from './components/app/app';
 import './global.css';
 
-const app = new App();
+const app: any = new App();
 app.start();
